+++ .eslintrc.js
@@ -4,8 +4,8 @@ module.exports = {
     es2021: true,
     node: true
   },
-  extends: [
-    'standard'
+  extends: 'standard',
+  overrides: [
   ],
   parserOptions: {
     ecmaVersion: 'latest'